refactor(anchor-navigation): remove style override props
BREAKING CHANGE:
Fixes FE-2758
