+++ src/components/anchor-navigation/anchor-navigation.d.ts
@@ -4,12 +4,6 @@ export interface AnchorNavigationProps {
   children?: React.ReactNode;
   /** The AnchorNavigationItems components to be rendered in the sticky navigation */
   stickyNavigation?: React.ReactNode;
-  /** Allows to override existing component styles */
-  styleOverride?: {
-    root?: () => object | object;
-    navigation?: () => object | object;
-    content?: () => object | object;
-  };
 }
 
 declare function AnchorNavigation(props: AnchorNavigationProps): JSX.Element;
