+++ src/components/anchor-navigation/anchor-navigation.stories.mdx
@@ -84,12 +84,7 @@ const MyComponent = () => (
   <Story id="test-anchornavigation--in-full-screen-dialog" name="InFullScreenDialog" />
 </Preview>
 
-### With overriden styles
-`AnchorNavigation`, `AnchorNavigationItem` and `AnchorSectionDivider` styles can be easily overriden by providing `styleOverride` prop with proper structure as defined in propTypes.
-
-<Story id="test-anchornavigation--with-overriden-styles" name="WithOverridenStyles" />
-
-## Props
+## Props:
 
 ### AnchorNavigation
