+++ src/components/anchor-navigation/index.d.ts
@@ -1,3 +1,2 @@
-export { default as AnchorNavigation } from "./anchor-navigation";
-export { default as AnchorNavigationItem } from "./anchor-navigation-item";
-export { default as AnchorSectionDivider } from "./anchor-section-divider";
+export { default as AnchorNavigation } from './anchor-navigation';
+export { default as AnchorNavigationItem } from './anchor-navigation-item';
