+++ src/components/anchor-navigation/index.js
@@ -1,3 +1,3 @@
 export { default as AnchorNavigation } from "./anchor-navigation.component";
 export { default as AnchorNavigationItem } from "./anchor-navigation-item.component";
-export { default as AnchorSectionDivider } from "./anchor-section-divider.component";
+export { StyledAnchorDivider as AnchorSectionDivider } from "./anchor-navigation.style";
